more dist removeal
