+++ conda/core/package_cache_data.py
@@ -209,7 +209,7 @@ class PackageCacheData(object):
                          for cache in cls.all_caches_writable_first() if cache), None)
         if pc_entry is not None:
             return pc_entry
-        raise CondaError("No package '%s' found in cache directories." % Dist(package_ref))
+        raise CondaError("No package '%s' found in cache directories." % package_ref.dist_str())
 
     @classmethod
     def tarball_file_in_cache(cls, tarball_path, md5sum=None, exclude_caches=()):
@@ -695,15 +695,3 @@ def rm_fetched(dist):
 def download(url, dst_path, session=None, md5=None, urlstxt=False, retries=3):
     from ..gateways.connection.download import download as gateway_download
     gateway_download(url, dst_path, md5)
-
-
-class package_cache(object):
-
-    def __contains__(self, dist):
-        return bool(PackageCacheData.first_writable().get(Dist(dist).to_package_ref(), None))
-
-    def keys(self):
-        return (Dist(v) for v in itervalues(PackageCacheData.first_writable()))
-
-    def __delitem__(self, dist):
-        PackageCacheData.first_writable().remove(Dist(dist).to_package_ref())
